+++ .mocharc.js
@@ -7,7 +7,7 @@ module.exports = {
     timeout: 999999,
     'preserve-symlinks': true,
     spec: [
-        './test/**/*.test.ts',
+        './test/**/*.test.ts'
     ],
     parallel: false
 }