+++ CHANGELOG.md
@@ -1,2 +1,9 @@
-### 1.2.0 (2019-11-15)
-- [NEW] IDBlock: 2Bit LengthType
+# CHANGELOG
+
+## [2.0.5-dev.0] - 2021-10-17
+- `CustomTypeSchema` supports
+- `customTypes` support for constructor options of `TSBuffer`
+
+## [1.2.0] - 2019-11-15
+### Added
+- IDBlock: 2Bit LengthType